Add files via upload
This if the second time I've worked with GitHub, forgive me for any mistakes I make I don't know how to use it.

In the video I show the steps I took to get the grasshopper script working with the provided sample.

In short the steps are as follow ;
Select substrate = part of model that doesn't require a supportless structure to print, really just a base part of the model on which the other parts are supportlessly printed.
Select Brep = The parts that are supportlessly printed on the substrate

Change layer number from 17 to something lower until the error goes away.
The path curve after gate will be fine once you enable slicing.
For the simulation you have to explode the cluster and find the path to the prusa file and replace this with your own path. Currently it points to Freddy's drive which only he has.
The final issue is not seeing the simulation, for this you have to enable "preview contents" of the Simulation cluster.